# frontend/public/openapi.yaml
openapi: 3.1.0
info:
  title: GenQMA Intelligence & Payments API
  description: '# GenQMA Intelligence & Payments API


    GenQMA is a provider marketplace for quantitative market intelligence. Clients
    discover live signals, select a provider-bound report, create a query-bound invoice,
    verify Circle Gateway x402 settlement, and retrieve the resulting preview or full
    report.


    ## Recommended buyer flow

    1. Discover providers with `GET /api/v1/providers` or live candidates with `GET
    /api/v1/agent/recommendations`.

    2. Optional: call `POST /api/v1/payment/quote` for a provider-bound price.

    3. Create an invoice with `POST /api/v1/payment/invoice`.

    4. Pay the returned Circle Gateway requirement, then verify with `POST /api/v1/payment/verify`.

    5. Poll `GET /api/v1/payment/invoices/{invoice_id}/status` when settlement is
    still processing.

    6. Retrieve the paid provider report from the provider-specific Reports route.


    ## Agent flow

    Use `POST /api/v1/agent/decision` to turn a bounded budget and provider/tier policy
    into a validated purchase decision. The payment executor remains responsible for
    signing and submitting payment; this API does not custody browser wallet keys.


    ## Access levels

    - **Public:** no credential is required. Optional-auth operations return additional
    private data only after verification.

    - **Protected:** requires a paid-access token, wallet-owner token, invoice secret,
...

...

...


...

...
servers:
...
paths:
...
  /api/v1/treasury/usyc/position:
    get:
      tags:
      - Agent decisioning
      summary: Get on-chain USYC position on Arc Testnet
      description: Retrieve live on-chain USYC balance, equivalent USDC assets, and
        APY.
      operationId: get_usyc_position_api_v1_treasury_usyc_position_get
      parameters:
      - name: account
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Account address to inspect
          title: Account
        description: Account address to inspect
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/USYCPositionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-qma-access: public
      x-qma-audiences:
      - agent
      - public
  /api/v1/treasury/usyc/sweep:
    post:
      tags:
      - Agent decisioning
      summary: Autonomous sweep of idle USDC cash into USYC
      description: Prepare and record an autonomous corporate treasury idle sweep
        into USYC yielding vault.
      operationId: sweep_idle_cash_api_v1_treasury_usyc_sweep_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SweepRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/USYCSweepResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-qma-access: public
      x-qma-audiences:
      - agent
      - public
  /api/v1/treasury/usyc/jit-redeem:
    post:
      tags:
      - Agent decisioning
      summary: Just-In-Time USYC redemption for bill payment
      description: Prepare and record JIT redemption of USYC into liquid USDC for
        x402 bills.
      operationId: jit_redeem_usyc_api_v1_treasury_usyc_jit_redeem_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JITRedeemRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/USYCJITRedeemResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-qma-access: public
      x-qma-audiences:
      - agent
      - public
  /api/v1/treasury/usyc/forecast:
    get:
      tags:
      - Agent decisioning
      summary: CFO cash-flow forecast and yield runway
      description: Predictive cash-flow forecasting and APY earnings projection for
        autonomous agents.
      operationId: get_treasury_forecast_api_v1_treasury_usyc_forecast_get
      parameters:
      - name: liquid_usdc
        in: query
        required: false
        schema:
          type: number
          minimum: 0
          default: 10.0
          title: Liquid Usdc
      - name: usyc_assets
        in: query
        required: false
        schema:
          type: number
          minimum: 0
          default: 100.0
          title: Usyc Assets
      - name: upcoming_bills_usdc
        in: query
        required: false
        schema:
          type: number
          minimum: 0
          default: 5.0
          title: Upcoming Bills Usdc
      - name: horizon_days
        in: query
        required: false
        schema:
          type: integer
          maximum: 365
          minimum: 1
          default: 30
          title: Horizon Days
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/USYCForecastResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-qma-access: public
      x-qma-audiences:
      - agent
      - public
  /api/v1/treasury/audit/euthyna:
    get:
      tags:
      - Agent decisioning
      summary: Get continuous Euthyna audit trail
      description: Retrieve immutable audit records for regulatory and board examination.
      operationId: get_audit_trail_api_v1_treasury_audit_euthyna_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 50
          title: Limit
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
      - name: actor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Actor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EuthynaAuditRecordResponse'
                title: Response Get Audit Trail Api V1 Treasury Audit Euthyna Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-qma-access: public
      x-qma-audiences:
      - agent
      - public
...
components:
  schemas:
...
    EuthynaAuditRecordResponse:
      properties:
        record_id:
          type: string
          title: Record Id
          description: Unique immutable audit record ID
        timestamp:
          type: string
          title: Timestamp
          description: ISO 8601 UTC timestamp
        action:
          type: string
          title: Action
          description: Action type (IDLE_SWEEP, JIT_REDEMPTION, X402_PAYMENT, etc.)
        actor:
          type: string
          title: Actor
          description: Agent or treasury address
        amount_usdc:
          type: number
          title: Amount Usdc
          description: Transaction amount in USDC
        treasury_liquid_before:
          type: number
          title: Treasury Liquid Before
          description: Liquid USDC before action
        treasury_liquid_after:
          type: number
          title: Treasury Liquid After
          description: Liquid USDC after action
        usyc_vault_shares:
          type: number
          title: Usyc Vault Shares
          description: USYC shares held in vault
        tx_hash:
          type: string
          title: Tx Hash
          description: On-chain Arc transaction hash or simulation hash
        arcscan_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Arcscan Url
          description: Arcscan explorer link
        policy_rule_applied:
          type: string
          title: Policy Rule Applied
          description: Corporate treasury rule applied
        cfo_reasoning:
          type: string
          title: Cfo Reasoning
          description: AI CFO rationale
        provider_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider Id
          description: Provider ID if applicable
        genlayer_consensus:
          anyOf:
          - type: string
          - type: 'null'
          title: Genlayer Consensus
          description: GenLayer multi-validator verdict
        integrity_hash:
          type: string
          title: Integrity Hash
          description: SHA-256 integrity digest
        status:
          type: string
          title: Status
          description: Verification status
          default: VERIFIED_AUDITABLE
      type: object
      required:
      - record_id
      - timestamp
      - action
      - actor
      - amount_usdc
      - treasury_liquid_before
      - treasury_liquid_after
      - usyc_vault_shares
      - tx_hash
      - policy_rule_applied
      - cfo_reasoning
      - integrity_hash
      title: EuthynaAuditRecordResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    JITRedeemRequest:
      properties:
        amount_usdc_needed:
          type: number
          exclusiveMinimum: 0.0
          title: Amount Usdc Needed
          description: USDC amount required for bill settlement
        receiver:
          anyOf:
          - type: string
          - type: 'null'
          title: Receiver
          description: Receiver of redeemed USDC
        owner:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner
          description: Owner of USYC shares to redeem
        cfo_reasoning:
          anyOf:
          - type: string
          - type: 'null'
          title: Cfo Reasoning
          description: Autonomous AI CFO rationale for JIT redemption
      type: object
      required:
      - amount_usdc_needed
      title: JITRedeemRequest
...
    SweepRequest:
      properties:
        amount_usdc:
          type: number
          exclusiveMinimum: 0.0
          title: Amount Usdc
          description: Amount of idle USDC to sweep into USYC
        depositor:
          anyOf:
          - type: string
          - type: 'null'
          title: Depositor
          description: Depositor address (defaults to platform treasury)
        cfo_reasoning:
          anyOf:
          - type: string
          - type: 'null'
          title: Cfo Reasoning
          description: Autonomous AI CFO rationale for this sweep
      type: object
      required:
      - amount_usdc
      title: SweepRequest
...
    USYCForecastResponse:
      properties:
        horizon_days:
          type: integer
          title: Horizon Days
          description: Forecast period in days
        current_liquid_usdc:
          type: number
          title: Current Liquid Usdc
          description: Current liquid USDC balance
        current_usyc_usdc:
          type: number
          title: Current Usyc Usdc
          description: Current USYC assets in USDC
        total_treasury_usdc:
          type: number
          title: Total Treasury Usdc
          description: Total treasury assets
        upcoming_bills_usdc:
          type: number
          title: Upcoming Bills Usdc
          description: Upcoming bills due
        projected_yield_earned_usdc:
          type: number
          title: Projected Yield Earned Usdc
          description: Projected yield earned over horizon
        effective_apy:
          type: string
          title: Effective Apy
          description: Effective APY string
          default: 5.0%
        action_recommended:
          type: string
          title: Action Recommended
          description: CFO recommendation (SWEEP_IDLE, REDEEM_JIT, etc.)
        safety_buffer_ratio:
          type: number
          title: Safety Buffer Ratio
          description: Ratio of total treasury to upcoming obligations
      type: object
      required:
      - horizon_days
      - current_liquid_usdc
      - current_usyc_usdc
      - total_treasury_usdc
      - upcoming_bills_usdc
      - projected_yield_earned_usdc
      - action_recommended
      - safety_buffer_ratio
      title: USYCForecastResponse
    USYCJITRedeemResponse:
      properties:
        status:
          type: string
          title: Status
          description: Redemption intent status
          default: PREPARED
        intent:
          additionalProperties: true
          type: object
          title: Intent
          description: Redemption call intent and parameters
        audit_record:
          $ref: '#/components/schemas/EuthynaAuditRecordResponse'
          description: Audit record for this action
      type: object
      required:
      - intent
      - audit_record
      title: USYCJITRedeemResponse
    USYCPositionResponse:
      properties:
        account:
          type: string
          title: Account
          description: Target treasury or agent account address
        vault_contract:
          type: string
          title: Vault Contract
          description: Deployed USYC vault contract on Arc Testnet
        underlying_asset:
          type: string
          title: Underlying Asset
          description: Underlying Arc Testnet USDC ERC-20 address
        chain_id:
          type: integer
          title: Chain Id
          description: Arc Testnet Chain ID
          default: 5042002
        network:
          type: string
          title: Network
          description: Network name
          default: Arc Testnet
        is_live_onchain:
          type: boolean
          title: Is Live Onchain
          description: Whether position was verified via live Arc RPC
        usyc_shares:
          type: number
          title: Usyc Shares
          description: USYC share balance
        usdc_equivalent:
          type: number
          title: Usdc Equivalent
          description: Underlying USDC value
        total_vault_assets_usdc:
          type: number
          title: Total Vault Assets Usdc
          description: Total vault assets across all depositors
        current_apy_percent:
          type: number
          title: Current Apy Percent
          description: Annualized percentage yield
          default: 5.0
        explorer_url:
          type: string
          title: Explorer Url
          description: Arcscan block explorer URL
      type: object
      required:
      - account
      - vault_contract
      - underlying_asset
      - is_live_onchain
      - usyc_shares
      - usdc_equivalent
      - total_vault_assets_usdc
      - explorer_url
      title: USYCPositionResponse
    USYCSweepResponse:
      properties:
        status:
          type: string
          title: Status
          description: Sweep intent status
          default: PREPARED
        intent:
          additionalProperties: true
          type: object
          title: Intent
          description: Deposit call intent and parameters
        audit_record:
          $ref: '#/components/schemas/EuthynaAuditRecordResponse'
          description: Audit record for this action
      type: object
      required:
      - intent
      - audit_record
      title: USYCSweepResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    x-qma-admin-token:
      type: apiKey
      description: 'Static administrative token for platform operations.


        **Enforcement:**

        - **Required:** `POST /api/v1/providers/{provider_id}/toggle` and `POST /api/v1/creators/applications/{application_id}/review`.

        - **Required for an unfiltered application list:** `GET /api/v1/creators/applications`
        accepts an admin token to list all applications; without it, callers must
        provide a wallet filter.

        - **Conditional:** provider discovery/stats require the token only when `include_disabled=true`.

        Invalid or missing credentials on a protected admin operation return `403
        Forbidden`.'
      in: header
      name: x-qma-admin-token
    X-QMA-Access-Token:
      type: apiKey
      description: 'Short-lived JWT granted after a successful invoice settlement.
        Used to unlock paid report endpoints.


        **Retrieval:** A token can be returned by payment verification/status after
        the invoice reaches the server-accepted paid state.

        **Enforcement:** Required on paid report endpoints and `POST /api/v1/chat`
        via the `X-QMA-Access-Token` header. Missing, invalid, expired, or invoice-mismatched
        tokens trigger `403 Forbidden`.'
      in: header
      name: X-QMA-Access-Token
    X-QMA-Wallet-Token:
      type: apiKey
      description: 'Wallet profile session JWT, generated by signing a SIWE-like message
        proving wallet ownership.


        **Retrieval:** Returned by the `/api/v1/wallets/{address}/session` endpoint.
...